backend/requirements.txt
# Python 3.10+ recommended (uvicorn / pydantic v2 need 3.9+).
#
# Local package: installs `app` into the venv so `uvicorn app.main:app` works
# regardless of current working directory (path is relative to this file).
-e .
#
# Web framework
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9,<1.0

# Auth: bcrypt only (passlib removed: bcrypt 4.1+ compat).
bcrypt>=4.0
python-jose[cryptography]>=3.3,<4.0

# Database(SQL)
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0

# Database(NoSQL)
pymongo>=4.6,<5.0

# Validation/settings
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0

# Migration (optional, recommended)
alembic>=1.13,<2.0

# AI
google-genai>=0.5,<1.0
huggingface_hub>=0.26
groq>=0.9

# HTTP Client
httpx>=0.27,<1.0

# Video (frame extract + optional YouTube URL download)
opencv-python-headless>=4.8,<5.0
# opencv-python          # optional: 로컬 GUI 필요 시 (서버는 headless 권장, 동시 설치 비권장)
yt-dlp>=2024.1

# (Optional: local HF / future extension)
transformers>=4.40
torch>=2.1

pydantic[email]
Pillow
requests
